models/NT/requirements-NT.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
argtools==0.1.6
async-timeout==4.0.3
attrs==23.2.0
biopython==1.83
boto3==1.34.19
botocore==1.34.19
cached-property==1.5.2
certifi @ file:///croot/certifi_1700501669400/work/certifi
charset-normalizer==3.3.2
chex==0.1.7
cmake==3.28.3
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
deepspeed==0.5.9
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-haiku==0.0.9
dm-tree==0.1.8
einops==0.7.0
etils==1.3.0
fair-esm @ git+https://github.com/facebookresearch/esm.git@2b369911bb5b4b0dda914521b9475cad1656b2ac
filelock==3.13.1
flax==0.7.2
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
hjson==3.1.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
jax==0.3.25
jaxlib==0.3.25+cuda11.cudnn82
Jinja2==3.1.3
jmespath==1.0.1
jmp==0.0.4
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.0
lit==18.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
nucleotide-transformer @ file:///zangzelin/duanchenrui/nucleotide-transformer
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.3
packaging==23.2
pandas==2.0.3
pillow==10.2.0
psutil==5.9.7
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorstore==0.1.45
termcolor==2.3.0
threadpoolctl==3.2.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.3.0
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
xgboost==2.0.3
yarl==1.9.4
zipp==3.17.0
